requirements.txt
pyyaml
jinja2
ConfigParser
termcolor
